Adding support for ZSH
It seems to be sufficient to update the color template to get this to work on
ZSH. Added a command line flag 'shell' instead of trying to detect the current
shell, which is always problematic. The powerline object also remembers this
argument, so you can customize based for each shell.