Moved all repo stats code from git.py to the base
When all of the new code to provide more interesting stats was added to
git.py, it was not very re-usable. A few pull requests could use it
though. In particular, #105 adds more info to the svn segment and #210
does the same for mercurial. I would rather not create inconsistencies
among these segments or have them duplicate code. Moving it into the
base file creates a place where each segment can access it and have a
consistent behavior.