Real Virtual env name while using "pipenv". pipenv is an officially recomended tool for managing python packages dependencies. In pipenv we have an option to create virtual env inside the project it self. The created virutal env name is .venv.
The normal bash display the name of the virual env as the parent folder
name.
Ex: new_shell/.venv then the virual env name displayed in base is
new_shell.

In powerline_shell the virtual env name is displayed as .venv.

This PR is a fix for displaying the parent folder as virtual env name.