Optimize dark theme colorize for hostname
When hostname colorization is enabled, then a
random number is being generated for the foreground
color of the hostname and the opposite color is
being calculated for the background color.

When running in a shell with dark background color
it typically looks better if the background color
of the hostname is "bright". So I introduced an
option in the configuration named "hostname.dark"
that (when set to true) picks the brighter of the
two colors for the hostname as the background color.