feature/truecolor: TrueColor Theming Support
Adds the ability to use (R,G,B) tuples in theme files to more accurately
specify the colors desired for each segment.
Not all terminals have TrueColor support, so in the case that a theme
uses (R,G,B) tuples that cannot be supported by the current terminal,
the colors are approximated using colortrans.rgb2short to bring them to
the nearest xterm-256 color.